Add add_cmdline_args() to DialogPartnerWorld
