+++ parlai/core/worlds.py
@@ -39,17 +39,18 @@ All worlds are initialized with the following parameters:
 import copy
 import random
 
-from typing import List, Dict, Union
+from typing import Dict, List, Optional, Union
 
+import parlai.utils.logging as logging
 from parlai.core.agents import create_agents_from_shared
 from parlai.core.loader import load_task_module, load_world_module
 from parlai.core.metrics import aggregate_named_reports
 from parlai.core.opt import Opt
+from parlai.core.params import ParlaiParser
 from parlai.core.teachers import Teacher, create_task_agent_from_taskname
 from parlai.utils.data import DatatypeHelper
 from parlai.utils.misc import Timer, display_messages
 from parlai.tasks.tasks import ids_to_tasks
-import parlai.utils.logging as logging
 
 
 def validate(observation):
@@ -303,6 +304,17 @@ class DialogPartnerWorld(World):
     This basic world switches back and forth between two agents, giving each agent one
     chance to speak per turn and passing that back to the other one.
     """
+
+    @classmethod
+    def add_cmdline_args(
+        cls, parser: ParlaiParser, partial_opt: Optional[Opt] = None
+    ) -> ParlaiParser:
+        """
+        Return the parser as-is.
+
+        Self-chat-specific world flags can be added here.
+        """
+        return parser
 
     def __init__(self, opt: Opt, agents=None, shared=None):
         if not ((agents is not None) ^ (shared is not None)):
